Fraud Detection In Banks Using ARIMA and  ML/Deep Learning Models
The goal and purpose of the project are to detect fraudulent behaviour in the bank to prevent associated risks and financial fraud. Neo-banks use ML to detect fraud and malicious attempts. The fraud detection system can be built using time series techniques integrated with other banking applications, such as credit risk scoring or online banking fraud. ARIMA combined with recurrent neural networks can be suitable to work on such an application.